Move `MouseHint` into `mouse.h`
